service_backend/requirements/dev.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

# Testing
pytest==6.2.2
pytest-flask==1.2.0
pytest-xdist==2.2.1
WebTest==2.0.35
factory-boy==3.2.0
pdbpp==0.10.2

# Lint and code style
black==20.8b1
flake8==3.9.0
flake8-blind-except==0.2.0
flake8-debugger==4.0.0
flake8-docstrings==1.6.0
flake8-isort==4.0.0
isort==5.7.0
pep8-naming==0.11.1
